Rename PS4_completed2.Rmd to Randomized Control Trials.Rmd
Recreated parts of analysis of the research paper Underinvestment in a Profitable Technology: the Case of Seasonal Migration in Bangladesh that used RCTs in identifying the causal effect of providing incentives to migrate during seasonal droughts. visualised effects and relationships using regression analysis with fixed effects and through ggplot libraries.